1. Показываем метки дисков из реестра при отсутствии физических. 2. "Починим" загрузку плагинов c нетрадиционным расположением зависимых dll. 3. Не закрываем quick search при потере фокуса, если OC новее, чем XP/2k3. 4. Рефакторинг.

+++ far/dlgedit.cpp
@@ -332,7 +332,7 @@ void DlgEdit::SetString(const string& Str, bool disable_autocomplete, int pos)
 		if (disable_autocomplete && (acompl = lineEdit->GetAutocomplete()))
 			lineEdit->SetAutocomplete(false);
 
-		lineEdit->SetString(Str.data(),Str.size());
+		lineEdit->SetString(Str.data(), static_cast<int>(Str.size()));
 		if (pos >= 0)
 			lineEdit->SetCurPos(pos);
